wait for led_classdevs to be created
This adds a hacky sleep prior to creating the phys_ctlrs. This gives the
driver time to create the led_classdevs (they are created after the
input devices). A better solution would be to have another udev rule
when the led_classdevs are created, but that would be a lot more work.