uinput: ff: fix rumble for dual Joycon and Snes
This fixes an issue where every ff effect requested to allocate a new one on the device.
That was making every effect to be the first one and then deplete the maximum allowed effects on the device.

So, actually use the rumble effects map for checking if already uploaded instead of incorrectly warning and readding the same effect as a new one.

Context: FF uapi in linux expects an id of -1 for allocating a new effect and an existing one for updating it.